+++ zenodot-tests/test/dd/kms/zenodot/tests/evaluationTests/VariableTest.java
@@ -22,25 +22,29 @@ public class VariableTest extends EvaluationTest
 		EvaluationTestBuilder testBuilder = new EvaluationTestBuilder()
 			.testInstance(testInstance)
 			.configurator(test -> {
-				test.createVariable("xyz", 15.0);
-				test.createVariable("abc", "Test");
+				test.createVariable("xyz", 15.0, true);
+				test.createVariable("abc", "Test", true);
+				test.createVariable("modifiableVariable", 42, false);
 			});
 
 		testBuilder
-			.addTest("b + xyz",			18.0)
-			.addTest("xyz * i",			-15000.0)
-			.addTest("(int) xyz / f",	6.0f)
-			.addTest("b + abc",			"3Test")
-			.addTest("abc + i",			"Test-1000")
-			.addTest("abc + f",			"Test2.5")
-			.addTest("xyz + xyz",		30.0)
-			.addTest("abc + abc",		"TestTest")
-			.addTest("test(xyz)",		"15.0")
-			.addTest("test(abc)",		"Test");
+			.addTest("b + xyz",											18.0)
+			.addTest("xyz * i",											-15000.0)
+			.addTest("(int) xyz / f",									6.0f)
+			.addTest("b + abc",											"3Test")
+			.addTest("abc + i",											"Test-1000")
+			.addTest("abc + f",											"Test2.5")
+			.addTest("xyz + xyz",										30.0)
+			.addTest("abc + abc",										"TestTest")
+			.addTest("test(xyz)",										"15.0")
+			.addTest("test(abc)",										"Test")
+			.addTest("modifiableVariable",								42)
+			.addTest("(modifiableVariable = 27) + modifiableVariable",	54);
 
 		testBuilder
 			.addTestWithError("xyz = 13")
-			.addTestWithError("abc = \"Test\"");
+			.addTestWithError("abc = \"Test\"")
+			.addTestWithError("modifiableVariable = \"Test\"");
 
 		return testBuilder.build();
 	}